requirements.txt
pandas>=2.2.2
python-dotenv>=1.0.0,<1.1.0
xlsxwriter>=3.0.9,<3.1.0
openpyxl>=3.1.2,<3.2.0
sqlalchemy>=2.0.4,<2.1.0
requests>=2.28.2,<2.29.0
fake-useragent>=1.1.3,<1.2.0
tonsdk>=1.0.12,<1.1.0
pyTelegramBotAPI>=4.11.0,<4.12.0
psycopg2-binary==2.9.9
asyncpg==0.29.0